

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.r
\name{CRMiss}
\alias{CRMiss}



\title{Abundance estimation from capture-recapture data when covariates are missing at random}



\usage{

CRMiss(cov.bin = NULL, cov.miss, cap.times, ind.obs, 
K, level = 0.95, M = 100)

}



\arguments{

\item{cov.bin}{a vector, the completely observed binary covariate.}

\item{cov.miss}{a vector, the covariate which is subject to missing.}

\item{cap.times}{a vector, where each element means the number of times of being captured 
for captured individuals.}

\item{ind.obs}{a vector, where elements take 0 or 1. 
1 means \code{cov.miss} is not missing and 0 otherwise.}

\item{K}{the number of capture occasions.}

\item{level}{the nominal confidence level.}

\item{M}{the number of imputed datasets when using the multiple imputation method.}

}



\value{

A list containing the following components:
    
\item{num.cap}{the number of captured individuals.}

\item{num.obs}{the number of captured individuals without missing covariates.}

\item{nu.el}{the maximum empirical likelihood estimate of the abundance.}

\item{ci.el}{the empirical likelihood ratio confidence interval of the abundance.}

\item{beta.el}{the maximum empirical likelihood estimate of parameters in the Alho-Huggin's model.}

\item{gamma0.el}{the maximum empirical likelihood estimate of the probabilities of being never captured.}

\item{gammajk.el}{the maximum empirical likelihood estimate of the probabilities of being captured k times 
and \code{cov.bin=j} in the extension case. }

\item{loglike.el}{the value of the maximum profile empirical log-likelihood.}

\item{num.para}{the number of parameters in the partial likelihood.}

\item{converge}{the convergence indicator of the algorithm when maximizing the empirical likelihood.
It takes 1 if the algorithm converges and 0 otherwise.}

\item{nu.el.cc}{the naive maximum empirical likelihood estimate of the abundance based on the complete cases.}

\item{ci.el.cc}{the naive empirical likelihood ratio confidence interval of the abundance based on the complete cases.}

\item{beta.el.cc}{the naive maximum empirical likelihood estimate of parameters in the Alho-Huggin's model based on the complete cases.}

\item{nu.ipw}{the estimate of the abundance based on the inverse probability weighting method.}

\item{sigma2.nu.ipw}{the variance estimate of \code{nu.ipw}}

\item{beta.ipw}{the estimates of parameters in the Alho-Huggin's model based on the inverse probability weighting method.}

\item{nu.mi}{the estimate of the abundance based on the multiple imputation method.}

\item{sigma2.nu.mi}{the variance estimate of \code{nu.mi}}

\item{beta.mi}{the estimates of parameters in the Alho-Huggin's model based on the multiple imputation method.}


}



\description{Four estimation methods are carried out under capture-recapture models when covariates are missing at random.}



\details{
When covariates are missing at random, four estimation methods are carried out under capture-recapture models.

First method is based on a new maximum empirical likelihood.
Refer to Liu et al. (2019+).

Second method is based on the naive maximum empirical likelihood derived from the complete cases.
See Liu et al. (2017).

The last two are based on the inverse probability weighting and multiple imputation methods.
See Lee et al. (2016).
}




\references{

Lee, S.-M., W.-H. Hwang, and J. de Dieu Tapsoba (2016). 
Estimation in closed capture–recapture models when covariates are missing at random. 
\emph{Biometrics} \strong{72}(4), 1294--1304.

Liu, Y., P. Li, and J. Qin (2017). 
Maximum empirical likelihood estimation for abundance in a closed
population from capture-recapture data. 
\emph{Biometrika} \strong{104}(3), 527--543.

Liu, Y., Y. Liu, P. Li, and L. Zhu (2019+).
Maximum likelihood abundance estimation from capture-recapture data when covariates are missing at random.
Submitted.
}




\examples{
## Analysis results of the prinia data set.

## Model 1 
outcomes.model1 <- CRMiss(cov.bin = prinia$fat.index, cov.miss = prinia$tail.length, cap.times = prinia$number.of.capture, 
                          ind.obs = (prinia$tail.length != 0) + 0, K = 17, level = 0.95, M = 100)

## Model 2
outcomes.model2 <- CRMiss(cov.miss = prinia$tail.length, cap.times = prinia$number.of.capture, 
                       ind.obs = (prinia$tail.length != 0) + 0, K = 17, level = 0.95, M = 100)

}









